requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
google-auth==2.35.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.1
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
prompt_toolkit==3.0.47
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyfcm==2.0.7
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
requests==2.32.3
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.3
zope.interface==7.0.3
